BLE and Spotify working

+++ src/app/services/spotify-api.service.ts
@@ -1,6 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
 import { Constants } from '../services/constants';
+import { Observable } from 'rxjs';
+import { Subject } from 'rxjs';
+import { BluetoothService } from './bluetooth.service';
 declare var cordova;
 
 interface Playlist {
@@ -45,23 +48,34 @@ export class SpotifyApiService {
 
   authToken: any;
 
+  currentTrackPosition: Observable<any>;
+  trackChange = new Subject<Track>();
+  currentTrack: Track;
+
+  match = false;
+  maxBpm = 220;
+  minBpm = 50;
+
   tracks = [];
 
   constructor(
-    private http: HttpClient
+    private http: HttpClient,
+    private bleService: BluetoothService
   ) { }
 
   initConnect(): Promise<any> {
-    return new Promise((resolve, reject) => {
-      localStorage.clear();
-      cordova.plugins.spotifyAuth.authorize(this.config).then(
-        (accessToken, expiresAt) => {
-          this.authToken = accessToken.accessToken;
-          console.log('Got an access token expiring at ' + expiresAt + ' :');
-          console.log(accessToken);
-          resolve(accessToken);
+    if (!this.authToken) {
+      return new Promise((resolve, reject) => {
+        localStorage.clear();
+        cordova.plugins.spotifyAuth.authorize(this.config).then(
+          (accessToken, expiresAt) => {
+            this.authToken = accessToken.accessToken;
+            console.log('Got an access token expiring at ' + expiresAt + ' :');
+            console.log(accessToken);
+            resolve(accessToken);
+        });
       });
-    });
+    }
   }
 
   async getAuth(): Promise<any> {
@@ -98,7 +112,7 @@ export class SpotifyApiService {
 
     const requestOptions = {
       headers: new HttpHeaders({
-        'Authorization': 'Bearer ' + this.authToken.access_token
+        'Authorization': 'Bearer ' + this.authToken
       })
     };
 
@@ -128,6 +142,114 @@ export class SpotifyApiService {
       this.tracks.push(track);
     }
     this.tracks.sort((a, b) => (a.tempo - b.tempo));
+    console.log('Spotify API : Tracks sorted array is :');
+    console.log(this.tracks);
   }
 
+  playTrack(uri): Promise<any> {
+    return new Promise((resolve, reject) => {
+      cordova.plugins.spotify.play(uri, {
+        clientId: this.clientId,
+        token: this.authToken
+      }).then(() => {
+        console.log('Spotify API : Track is playing ');
+        console.log(this.currentTrack);
+        this.currentTrackPosition = new Observable((observer) => {
+          setInterval(() => {
+            cordova.plugins.spotify.getPosition().then(pos => {
+              observer.next(pos);
+            });
+          }, 500);
+        });
+        setInterval(() => {
+          this.checkPosition();
+        }, 500);
+        this.trackChange.next(this.currentTrack);
+        resolve(this.currentTrack);
+
+      }).catch(error => {
+        reject(error);
+      });
+    });
+  }
+
+  playRandomTrack(): Promise<any> {
+    return new Promise((resolve, reject) => {
+      this.currentTrack = this.tracks[Math.floor(Math.random() * this.tracks.length)];
+      console.log('Spotify API : Random track is ');
+      console.log(this.currentTrack);
+      this.playTrack(this.currentTrack.uri).then(track => {
+        resolve(track);
+      }).catch(error => {
+        reject(error);
+      });
+    });
+  }
+
+  async checkPosition() {
+    const position = await cordova.plugins.spotify.getPosition();
+
+    if ((this.currentTrack.duration_ms - position) < 60000 && this.match === false) {
+      console.log('Spotify API : Starting record :');
+      this.match = true;
+    }
+
+    if ((this.currentTrack.duration_ms - position) < 500) {
+      const mean = this.bleService.total / this.bleService.count;
+      this.matchTrack(mean);
+      console.log('Spotify API : Track Ended : ' + this.currentTrack.name);
+    }
+  }
+
+  matchTrack(mean) {
+    let selection = [];
+    const range = mean / (this.maxBpm - this.minBpm);
+    console.log('Spotify API : the range on tracks[] is : ' + range);
+    const scope = Math.round(this.tracks.length * range);
+    if (scope < 5) {
+      selection = this.tracks.slice(0, scope + 5);
+    }
+    if (scope > this.tracks.length) {
+      selection = this.tracks.slice(scope - 5, this.tracks.length);
+    } else {
+      selection = this.tracks.slice(scope - 5, scope + 5);
+    }
+
+    console.log('Spotify API : the selection is : ');
+    console.log(selection);
+
+    let nextTrack = selection[Math.floor(Math.random() * selection.length)];
+
+    if (nextTrack.id === this.currentTrack.id) {
+      while (nextTrack.id === this.currentTrack.id) {
+        nextTrack = selection[Math.floor(Math.random() * selection.length)];
+      }
+    }
+
+    console.log('Spotify API : the next track is : ' + nextTrack);
+
+    this.currentTrack = nextTrack;
+    this.playTrack(nextTrack.uri);
+  }
+
+  resumeTrack(): Promise<number> {
+
+    return new Promise((resolve, reject) => {
+      cordova.plugins.spotify.resume()
+      .then(() => {
+          console.log('Spotify API : music is resuming');
+          resolve(1);
+      });
+    });
+  }
+
+  pauseTrack(): Promise<number> {
+    return new Promise((resolve, reject) => {
+      cordova.plugins.spotify.pause()
+      .then(() => {
+          console.log('Spotify API : music is paused');
+          resolve(1);
+      });
+    });
+  }
 }
